refactor: ailaStreamingStatus subscriptions (#560)
