+++ apps/nextjs/src/stores/chatStore/selectors.ts
@@ -1,8 +1,8 @@
-import type { ChatStore } from ".";
+import type { ChatState } from ".";
 
 /**
  * @example const canAppend = useChatStore(canAppend);
  */
-export const canAppendSelector = (store: ChatStore) =>
+export const canAppendSelector = (store: ChatState) =>
   store.ailaStreamingStatus === "Idle" ||
   (store.ailaStreamingStatus === "Moderating" && !store.queuedUserAction);
