+++ apps/nextjs/src/stores/chatStore/stateActionFunctions/handleAilaStreamingStatusUpdated.ts
@@ -1,0 +1,14 @@
+import { aiLogger } from "@oakai/logger";
+
+import type { ChatSetter, ChatGetter, AilaStreamingStatus } from "../types";
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const log = aiLogger("chat:store");
+
+export const handleAilaStreamingStatusUpdated =
+  (set: ChatSetter, get: ChatGetter) =>
+  (streamingStatus: AilaStreamingStatus) => {
+    if (streamingStatus === "Idle" && get().queuedUserAction) {
+      void get().executeQueuedAction();
+    }
+  };
