+++ apps/nextjs/src/stores/chatStore/stateActionFunctions/handleStreamingFinished.ts
@@ -3,9 +3,4 @@ import type { ChatSetter, ChatGetter } from "../types";
 export const handleStreamingFinished =
   (set: ChatSetter, get: ChatGetter) => () => {
     get().scrollToBottom();
-
-    // TODO: commented while we have the same in handleSetMessages
-    // if (get().queuedUserAction) {
-    //   void get().executeQueuedAction();
-    // }
   };
