+++ apps/nextjs/src/stores/lessonPlanStore/index.ts
@@ -1,7 +1,3 @@
-import type {
-  LessonPlanKey,
-  LooseLessonPlan,
-} from "@oakai/aila/src/protocol/schema";
 import { aiLogger } from "@oakai/logger";
 import { createStore } from "zustand";
 
@@ -9,45 +5,22 @@ import type { LessonPlanTrackingContextProps } from "@/lib/analytics/lessonPlanT
 import type { TrpcUtils } from "@/utils/trpc";
 
 import type { GetStore } from "../AilaStoresProvider";
-import type { AiMessage } from "../chatStore/types";
 import { logStoreUpdates } from "../zustandHelpers";
 import { handleMessagesUpdated } from "./stateActionFunctions/handleMessagesUpdated";
 import { handleRefetch } from "./stateActionFunctions/handleRefetch";
 import { handleTrackingEvents } from "./stateActionFunctions/handleTrackingEvents";
+import type { LessonPlanState } from "./types";
+
+export * from "./types";
 
 const log = aiLogger("lessons:store");
-
-export type LessonPlanStore = {
-  id: string;
-  lessonPlan: LooseLessonPlan;
-  appliedPatchHashes: string[];
-  appliedPatchPaths: LessonPlanKey[];
-  sectionsToEdit: LessonPlanKey[];
-  iteration: number | undefined;
-  isAcceptingChanges: boolean;
-  numberOfStreamedCompleteParts: number;
-  // Used for lessonPlanTracking.onStreamFinished
-  lastLessonPlan: LooseLessonPlan;
-  isShared: boolean;
-  scrollToSection: LessonPlanKey | null;
-
-  // setters
-  setScrollToSection: (sectionKey: LessonPlanKey | null) => void;
-
-  // actions
-  messageStarted: () => void;
-  messagesUpdated: (messages: AiMessage[]) => void;
-  messageFinished: () => void;
-  refetch: () => Promise<void>;
-  resetStore: () => void;
-};
 
 const initialPerMessageState = {
   sectionsToEdit: [],
   appliedPatchHashes: [],
   appliedPatchPaths: [],
   numberOfStreamedCompleteParts: 0,
-} satisfies Partial<LessonPlanStore>;
+} satisfies Partial<LessonPlanState>;
 
 export const createLessonPlanStore = ({
   id,
@@ -60,9 +33,9 @@ export const createLessonPlanStore = ({
   trpcUtils: TrpcUtils;
   getStore: GetStore;
   lessonPlanTracking: LessonPlanTrackingContextProps;
-  initialValues?: Partial<LessonPlanStore>;
+  initialValues?: Partial<LessonPlanState>;
 }) => {
-  const lessonPlanStore = createStore<LessonPlanStore>((set, get) => ({
+  const lessonPlanStore = createStore<LessonPlanState>((set, get) => ({
     id,
     lessonPlan: {},
     iteration: undefined,
