+++ apps/nextjs/src/stores/moderationStore/index.ts
@@ -1,31 +1,16 @@
-import type { PersistedModerationBase } from "@oakai/core/src/utils/ailaModeration/moderationSchema";
-import type { Moderation } from "@prisma/client";
 import { create } from "zustand";
 
 import type { GetStore } from "@/stores/AilaStoresProvider";
 import type { TrpcUtils } from "@/utils/trpc";
 
 import { logStoreUpdates } from "../zustandHelpers";
+import { handleAilaStreamingStatusUpdated } from "./actionFunctions/handleAilaStreamingStatus";
 import { handleFetchModerations } from "./actionFunctions/handleFetchModeration";
 import { handleToxicModeration } from "./actionFunctions/handleToxicModeration";
 import { handleUpdateModerationState } from "./actionFunctions/handleUpdateModerationState";
+import type { ModerationState } from "./types";
 
-export type ModerationStore = {
-  id: string;
-  moderations: Moderation[] | [];
-  isModerationsLoading: boolean | null;
-  toxicInitialModeration: PersistedModerationBase | null;
-  toxicModeration: PersistedModerationBase | null;
-  lastModeration: PersistedModerationBase | null;
-
-  updateToxicModeration: (mod: PersistedModerationBase | null) => void;
-  setLastModeration: (mod: PersistedModerationBase | null) => void;
-  setIsModerationsLoading: (isModerationsLoading: boolean) => void;
-  updateModerationState: (mods?: Moderation[]) => void;
-  fetchModerations: () => Promise<void>;
-
-  clearModerations: () => void;
-};
+export * from "./types";
 
 export const createModerationStore = ({
   id,
@@ -34,11 +19,11 @@ export const createModerationStore = ({
   trpcUtils,
 }: {
   id: string;
-  initialValues?: Partial<ModerationStore>;
+  initialValues?: Partial<ModerationState>;
   getStore: GetStore;
   trpcUtils: TrpcUtils;
 }) => {
-  const moderationStore = create<ModerationStore>((set, get) => ({
+  const moderationStore = create<ModerationState>((set, get) => ({
     id,
     moderations: [],
     isModerationsLoading: null,
@@ -64,6 +49,7 @@ export const createModerationStore = ({
         lastModeration: null,
       });
     },
+    ailaStreamingStatusUpdated: handleAilaStreamingStatusUpdated(set, get),
     ...initialValues,
   }));
   logStoreUpdates(moderationStore, "moderation:store");
